Added actual parser.
The other one was unreliable and unrealistic. This one has an array of tokens. It will append a keyword to parsedCode when it detects one outside of a string (I am sort of a hobbyist programming language desgner so seeing that implementation of a parser hurt me. Don't merge it yet, it hasn't been tested for errors
 I'll do all that in the morning.